pyproject.toml
[tool.poetry]
name = "wgfinder"
version = "0.3"
description = ""
authors = ["Yannik Queisler <dev.yannik.queisler@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
beautifulsoup4 = "^4.12"
click = "^8.1.7"
requests = "^2.32"
openai = "^1.40"
pywhatkit = "5.4"
Flask = "^3.0.0"
backoff = "^2.2.1"
python-dotenv = "^1.1.1"

[tool.poetry.dev-dependencies]
pytest = "^8.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
